[TASK] Step 1: Create the bare minimum structure
Whilst developing a template, we strongly encourage the ProviderExtension pattern.

The pattern allows the developer to put all assets, templates and configuration into one
extension, which could be delivered to a multitude of developer, staging, test or even production
 environments.

Once the extension is loaded in the Backend, and activated through inclusion of the static TypoScript
template and all dependencies are satisfied, there's nothing stopping you from delivering the
awesome project.