[TASK] Step 4: Create a page layout
This is not the real page layout! This wires up different partials and breaks the page layout down in its respective parts.

As you see, the Page.html layout file renders a Header and a Footer Partial. Those will be added soon, and let us focus on the content rendering, rather then caring about the whole layout.