[TASK] Step 6: Add a 3 rows, 1column backend layout
When utilizing fluidpages for Page layouts (as we do), we define out BackendLayouts with pure fluid.
Like I did, you can render very complex layouts with many rows and columns which can inherit and can contain custom configuration in the Configuration Section.

Page templates are placed into Resources/Private/Templates/Page as per default. To explain it with another convention:
Place the Templates for the flux PageController in Templates/Page. Just as you are used to with any other extbase extension.