[TASK] Step 7: Add a partial for the page assets (js/css)
Assets such as custom JS or CSS-from files or inline-can be inserted wherever you want. EXT:vhs' AssetService will take care of it.
With Assets, you can insert JS and CSS in any order and have it sorted via the ``dependencies`` attribute.

As most of your CSS and JS will have to be loaded on any page (think of modernizr and the common CSS for example), it only makes sense to place them together. We'll do that later on in SharedResources